+++ actix-server/tests/test_server.rs
@@ -22,11 +22,14 @@ fn test_bind() {
 
     let h = thread::spawn(move || {
         let sys = actix_rt::System::new("test");
-        let srv = Server::build()
-            .disable_signals()
-            .bind("test", addr, move || fn_service(|_| ok::<_, ()>(())))
-            .unwrap()
-            .start();
+        let srv = sys.block_on(lazy(|_| {
+            Server::build()
+                .workers(1)
+                .disable_signals()
+                .bind("test", addr, move || fn_service(|_| ok::<_, ()>(())))
+                .unwrap()
+                .start()
+        }));
         let _ = tx.send((srv, actix_rt::System::current()));
         let _ = sys.run();
     });
@@ -46,12 +49,14 @@ fn test_listen() {
     let h = thread::spawn(move || {
         let sys = actix_rt::System::new("test");
         let lst = net::TcpListener::bind(addr).unwrap();
-        Server::build()
-            .disable_signals()
-            .workers(1)
-            .listen("test", lst, move || fn_service(|_| ok::<_, ()>(())))
-            .unwrap()
-            .start();
+        sys.block_on(lazy(|_| {
+            Server::build()
+                .disable_signals()
+                .workers(1)
+                .listen("test", lst, move || fn_service(|_| ok::<_, ()>(())))
+                .unwrap()
+                .start()
+        }));
         let _ = tx.send(actix_rt::System::current());
         let _ = sys.run();
     });
@@ -77,18 +82,20 @@ fn test_start() {
 
     let h = thread::spawn(move || {
         let sys = actix_rt::System::new("test");
-        let srv: Server = Server::build()
-            .backlog(100)
-            .disable_signals()
-            .bind("test", addr, move || {
-                fn_service(|io: TcpStream| async move {
-                    let mut f = Framed::new(io, BytesCodec);
-                    f.send(Bytes::from_static(b"test")).await.unwrap();
-                    Ok::<_, ()>(())
+        let srv = sys.block_on(lazy(|_| {
+            Server::build()
+                .backlog(100)
+                .disable_signals()
+                .bind("test", addr, move || {
+                    fn_service(|io: TcpStream| async move {
+                        let mut f = Framed::new(io, BytesCodec);
+                        f.send(Bytes::from_static(b"test")).await.unwrap();
+                        Ok::<_, ()>(())
+                    })
                 })
-            })
-            .unwrap()
-            .start();
+                .unwrap()
+                .start()
+        }));
 
         let _ = tx.send((srv, actix_rt::System::current()));
         let _ = sys.run();
@@ -144,28 +151,30 @@ fn test_configure() {
     let h = thread::spawn(move || {
         let num = num2.clone();
         let sys = actix_rt::System::new("test");
-        let srv = Server::build()
-            .disable_signals()
-            .configure(move |cfg| {
-                let num = num.clone();
-                let lst = net::TcpListener::bind(addr3).unwrap();
-                cfg.bind("addr1", addr1)
-                    .unwrap()
-                    .bind("addr2", addr2)
-                    .unwrap()
-                    .listen("addr3", lst)
-                    .apply(move |rt| {
-                        let num = num.clone();
-                        rt.service("addr1", fn_service(|_| ok::<_, ()>(())));
-                        rt.service("addr3", fn_service(|_| ok::<_, ()>(())));
-                        rt.on_start(lazy(move |_| {
-                            let _ = num.fetch_add(1, Relaxed);
-                        }))
-                    })
-            })
-            .unwrap()
-            .workers(1)
-            .start();
+        let srv = sys.block_on(lazy(|_| {
+            Server::build()
+                .disable_signals()
+                .configure(move |cfg| {
+                    let num = num.clone();
+                    let lst = net::TcpListener::bind(addr3).unwrap();
+                    cfg.bind("addr1", addr1)
+                        .unwrap()
+                        .bind("addr2", addr2)
+                        .unwrap()
+                        .listen("addr3", lst)
+                        .apply(move |rt| {
+                            let num = num.clone();
+                            rt.service("addr1", fn_service(|_| ok::<_, ()>(())));
+                            rt.service("addr3", fn_service(|_| ok::<_, ()>(())));
+                            rt.on_start(lazy(move |_| {
+                                let _ = num.fetch_add(1, Relaxed);
+                            }))
+                        })
+                })
+                .unwrap()
+                .workers(1)
+                .start()
+        }));
         let _ = tx.send((srv, actix_rt::System::current()));
         let _ = sys.run();
     });